requirements.txt
boto3==1.21.5
botocore==1.24.5
jmespath==0.10.0
PyJWT==2.3.0
python-dateutil==2.8.2
s3transfer==0.5.1
six==1.16.0
urllib3==1.26.8
requests==2.27.1
uuid==1.30